changed float to double for more accuracy for higher temperatures

+++ SW02/Aufgabe_1/src/main/java/hslu/oop/sw02/Temperatur.java
@@ -2,59 +2,59 @@ package hslu.oop.sw02;
 
 public class Temperatur {
 
-    float temperatureInCelsius;
+    double temperatureInCelsius;
 
     /**
      * Default Constructor
      * sets default temperature to 20 degree Celsius
      */
     public Temperatur() {
-        this.temperatureInCelsius = 20.0f;
+        this.temperatureInCelsius = 20.0;
     }
 
     /**
      * Constructor
-     * sets default temperature to given float value degree Celsius
+     * sets default temperature to given double value degree Celsius
      */
-    public Temperatur(float temperatureInCelsius) {
+    public Temperatur(double temperatureInCelsius) {
         this.temperatureInCelsius = temperatureInCelsius;
     }
 
-    public float getTemperatureInCelsius() {
+    public double getTemperatureInCelsius() {
         return temperatureInCelsius;
     }
 
-    public void setTemperatureInCelsius(float temperatureInCelsius) {
+    public void setTemperatureInCelsius(double temperatureInCelsius) {
         this.temperatureInCelsius = temperatureInCelsius;
     }
 
     /**
      * @return current temperature of class in Kelvin
      */
-    public float getTemperatureInKelvin(){
-        return temperatureInCelsius  + 273.15f;
+    public double getTemperatureInKelvin(){
+        return temperatureInCelsius  + 273.15;
     }
 
     /**
      * @return current temperature of class in Fahrenheit
      */
-    public float getTemperatureInFahrenheit(){
-        return temperatureInCelsius * 1.8f + 32f;
+    public double getTemperatureInFahrenheit(){
+        return temperatureInCelsius * 1.8f + 32;
     }
 
     /**
      * changes temperature relative to given value
      * @param temperatureInKelvin relative temperature change in Kelvin
      */
-    public void changeTemperatureInKelvin(float temperatureInKelvin){
-        this.temperatureInCelsius =- temperatureInKelvin - 273.15f;
+    public void changeTemperatureInKelvin(double temperatureInKelvin){
+        this.temperatureInCelsius =- temperatureInKelvin - 273.15;
     }
 
     /**
      * changes temperature relative to given value
      * @param temperatureInCelsius relative temperature change in Celsius
      */
-    public void changeTemperatureInCelsius(float temperatureInCelsius){
+    public void changeTemperatureInCelsius(double temperatureInCelsius){
         this.temperatureInCelsius =- temperatureInCelsius;
     }
